gcp-version/requirements.txt
certifi
chardet==3.0.4
Click==7.0
dash==2.0.0
dash-auth==1.3.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.0.0
dash-renderer==1.2.4
dnspython==1.16.0
Flask==1.1.1
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
google-api-python-client
google-auth
google-auth-httplib2
google-cloud
google-cloud-storage
future==0.18.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python
pandas
plotly==5.0.0
protobuf
requests==2.22.0
retrying==1.3.3
six==1.14.0
ua-parser==0.8.0
urllib3==1.25.7
Werkzeug==0.16.0




